+++ both/controllers/Config.jsx
@@ -96,10 +96,6 @@ var navFMStaff = [{
 }];
 
 var navSupplier = [{
-	path:"dashboard",		
-	label:"Dashboard",
-	icon:"fa fa-newspaper-o"
-},{
 	path:"requests",		
 	label:"Jobs",
 	icon:"fa fa-wrench"